re-structured to give way to void
