+++ src/pages/projContracts/parts/Preview/PreviewMenu/MenuContainer.tsx
@@ -3,16 +3,27 @@ import { MouseEvent, useState } from 'react';
 import MoreIcon from '@mui/icons-material/More';
 import { MenuVoidContract } from './MenuVoidContract';
 import { MenuExcel } from './MenuExcel';
+import { useFormikContext } from 'formik';
+import { TypeOfForm } from '../../../form';
+import { StartContract } from './startContract';
 
 export const MenuContainer = () => {
 
+  const { values } = useFormikContext<TypeOfForm>();
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const open = Boolean(anchorEl);
   const handleClick = (e: MouseEvent<HTMLButtonElement>) => {
     setAnchorEl(e.currentTarget);
   };
 
+  const {
+    envelopeStatus,
+  } = values;
+
   const handleClose = () => setAnchorEl(null);
+
+  const isWithContract = !!envelopeStatus;
+  const isVoidable = envelopeStatus === 'sent';
 
   return (
     <div>
@@ -36,7 +47,10 @@ export const MenuContainer = () => {
         }}
       >
         <MenuExcel />
-        <MenuVoidContract handleClose={handleClose} />
+        {!isWithContract &&  <StartContract />}
+
+        {isVoidable && <MenuVoidContract handleClose={handleClose} />}
+
       </Menu>
 
     </div>
