+++ src/pages/projContracts/parts/PreviewToolBar/index.ts
@@ -2,4 +2,4 @@ export * from './DownloadContract';
 export * from './EnvelopeStatus';
 export * from './PreviewToolBar';
 export * from './RefreshButton';
-export * from '../Preview/startContract';
+export * from '../Preview/PreviewMenu/startContract';
